Code Update

+++ app/Http/Controllers/Admin/UserController.php
@@ -9,6 +9,7 @@ use Carbon\Carbon;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Mail;
 
 class UserController extends Controller
@@ -42,7 +43,8 @@ class UserController extends Controller
         $Affected = null;
         $Affected = User::create([
             'email' => $validatedData['email'],
-            'name' => $request['name'],
+            'first_name' => $request['first_name'],
+            'last_name' => $request['last_name'],
             'phone' => $request['phone'],
             'address' => $request['address'],
             'password' => Hash::make($request['password']),
@@ -70,11 +72,22 @@ class UserController extends Controller
 
     public function update(Request $request)
     {
+        // Validate the form data
+        $validator = Validator::make($request->all(), [
+            'email' => 'required|email|unique:users,email,' . $request->user()->id,
+        ]);
+
+        if ($validator->fails()) {
+            return redirect()->route('admin.profile')->with('error-message', 'Email Address already exist');
+        }
+
         /* 1 - Profile Picture */
         $FileProfile = "";
         if (!empty($request->has('profile_image'))) {
             if ($request['old_profile_image'] != "") {
-                $Path = public_path('storage/users') . '/' . $request['old_profile_image'];
+                $temp = explode("/", $request['old_profile_image']);
+                $fileName = end($temp);
+                $Path = public_path('storage/users/') . $fileName;
                 if (file_exists($Path)) {
                     unlink($Path);
                 }
@@ -82,13 +95,17 @@ class UserController extends Controller
             $FileProfile = 'ProfileImage-' . Carbon::now()->format('Ymd-His') . '.' . $request->file('profile_image')->extension();
             $request->file('profile_image')->storeAs('public/users/', $FileProfile);
         } else {
-            $FileProfile = $request['old_profile_image'];
+            $temp = explode("/", $request['old_profile_image']);
+            $fileName = end($temp);
+            $FileProfile = $fileName;
         }
         
         DB::beginTransaction();
         $Affected = null;
         $Affected = User::where('id', $request->user_id)->update([
-            'name' => $request['name'],
+            'first_name' => $request['first_name'],
+            'last_name' => $request['last_name'],
+            'email' => $request['email'],
             'phone' => $request['phone'],
             'address' => $request['address'],
             'role' => '1',
